+++ src/Controller/UserController.php
@@ -79,10 +79,11 @@ class UserController extends AbstractController
 
     #[Route('/delete/{id<\d+>}', name: 'app_user_delete')]
     #[IsGranted('ROLE_USER', message: 'Vous devez être connecté en tant qu\'utilisateu-rice pour accéder à cette page')]
-    public function delete(User $user, EntityManagerInterface $em): Response
+    public function delete(User $user, EntityManagerInterface $em, Request $request): Response
     {
         $user = $this->getUser();
-        if ($user) {
+        $submittedToken = $request->request->get('token');
+        if ($user && $this->isCsrfTokenValid('delete-account', $submittedToken)) {
             $em->remove($user);
             $em->flush();
             $this->addFlash('deleteAccount', 'Votre compte a été supprimé avec succès');
